requirements.txt
Django==3.1.14
djangorestframework==3.13.1
psycopg2>=2.8
PyJWT==1.7.1

